Move read/write config to utils
